requirements.txt
flask
pymongo
flask-login
gunicorn
werkzeug
Pillow
dnspython